Minor typo and missing assignment of Read 1
Hi, I just want to say that I really enjoyed reading the paper you wrote for this program, thank you for writing it so clearly and with a great focus on detail!

I have downloaded and run the program and came across a few types, noted above. 
Line 69: Space required between the "[" and "-z" statement, bash error if there is no space.
line 73-78: else statement doesn't assign Read1 to $R1 in the case of two reads provided. Causes errors further down with filtering and mapping 

All the best,
Andy